Task
A group of initiated mathematicians are interested in testing in a coded virtual environment the operation of an equation that is estimated to be able to solve dilemmas of current physics. That said, we asked a team of programmers from the Mtz university to implement the function in the JAVA programming language.
The equation will be provided in a JPEG file next to this Text Document